fixed the test failure caused by two variables declared as int instead of long, which caused jvm crash when offset overflowed to long

+++ core/src/main/scala/org/apache/spark/sql/execution/aggregate/SnappyHashAggregateExec.scala
@@ -617,7 +617,7 @@ case class SnappyHashAggregateExec(
     ctx.addMutableState("int", mapCounter, s"$mapCounter = 0;")
 
     val iterValueOffsetTerm = ctx.freshName("iterValueOffsetTerm")
-    ctx.addMutableState("int", iterValueOffsetTerm, s"$iterValueOffsetTerm = 0;")
+    ctx.addMutableState("long", iterValueOffsetTerm, s"$iterValueOffsetTerm = 0;")
 
     val nullKeysBitsetTerm = ctx.freshName("nullKeysBitset")
     val nullAggsBitsetTerm = ctx.freshName("nullAggsBitset")
